requirement.txt
django
crispy-bootstrap5==0.6
django-crispy-forms==1.14.0
django-filter
django-htmx
django-tables2